# backend/internal/modules/identity/adapters/postgres/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/shopspring/decimal"
)

type Category struct {
	ID        uuid.UUID
	TenantID  uuid.UUID
	Name      string
	ParentID  pgtype.UUID
	CreatedAt time.Time
}

type InventorySetting struct {
	TenantID           uuid.UUID
	AllowNegativeStock bool
}

type Location struct {
	ID        uuid.UUID
	TenantID  uuid.UUID
	Name      string
	Active    bool
	CreatedAt time.Time
}

type Product struct {
	ID          uuid.UUID
	TenantID    uuid.UUID
	Sku         string
	Barcode     *string
	Name        string
	SearchName  string
	Description string
	CategoryID  pgtype.UUID
	BaseUnit    string
	Cost        decimal.Decimal
	Price       decimal.Decimal
	TaxRate     decimal.Decimal
	Active      bool
	CreatedAt   time.Time
}

type RefreshToken struct {
	ID        uuid.UUID
	TenantID  uuid.UUID
	UserID    uuid.UUID
	FamilyID  uuid.UUID
	TokenHash []byte
	ExpiresAt time.Time
	CreatedAt time.Time
	RevokedAt *time.Time
}

type StockLevel struct {
	TenantID    uuid.UUID
	ProductID   uuid.UUID
	LocationID  uuid.UUID
	Quantity    decimal.Decimal
	AverageCost decimal.Decimal
	UpdatedAt   time.Time
}

type StockMovement struct {
	ID               uuid.UUID
	TenantID         uuid.UUID
	ProductID        uuid.UUID
	LocationID       uuid.UUID
	Type             string
	Quantity         decimal.Decimal
	UnitCost         decimal.Decimal
	BalanceAfter     decimal.Decimal
	AverageCostAfter decimal.Decimal
	Reason           string
	ReferenceKind    *string
	ReferenceID      pgtype.UUID
	UserID           uuid.UUID
	OccurredAt       time.Time
}

type Tenant struct {
	ID        uuid.UUID
	Name      string
	Nit       string
	Status    string
	CreatedAt time.Time
}

type User struct {
	ID           uuid.UUID
	TenantID     uuid.UUID
	Email        string
	Name         string
	PasswordHash string
	Role         string
	Active       bool
	CreatedAt    time.Time
}